fixed suggested changes
