+++ src/graphql/queries/getProcedureUpdates.js
@@ -1,8 +1,8 @@
 import gql from 'graphql-tag';
 
 export default gql`
-  query procedureUpdates($pageSize: Int, $period: [Int!], $type: [String!]) {
-    procedureUpdates(pageSize: $pageSize, period: $period, type: $type) {
+  query procedureUpdates($period: [Int!], $type: [String!]) {
+    procedureUpdates(period: $period, type: $type) {
       procedureId
       updatedAt
     }
