+++ src/graphql/queries/getProcedures.js
@@ -1,8 +1,8 @@
 import gql from 'graphql-tag';
 
 export default gql`
-  query procedures($pageSize: Int, $IDs: [String!]) {
-    procedures(pageSize: $pageSize, IDs: $IDs) {
+  query procedures($IDs: [String!]) {
+    procedures(IDs: $IDs) {
       title
       procedureId
       type
