+++ src/scripts/webhook.js
@@ -1,0 +1,56 @@
+import importProcedures from './import';
+import getProcedureUpdates from '../graphql/queries/getProcedureUpdates';
+import client from '../graphql/client';
+import ProcedureModel from '../models/Procedure';
+
+export default async (data) => {
+  // Count local Data in groups
+  const groups = await ProcedureModel.aggregate([{
+    // Group by Period & Type
+    $group: {
+      _id: { period: '$period', type: '$type' },
+      count: { $sum: 1 },
+    },
+  },
+  {
+    // Group by Period
+    $group: {
+      _id: '$_id.period',
+      types: { $push: { type: '$_id.type', count: '$count' } },
+    },
+  },
+  {
+    // Rename _id Field to period
+    $project: { _id: 0, period: '$_id', types: 1 },
+  }]);
+
+  const update = [];
+  await Promise.all(data.map(async (d) => {
+    const period = parseInt(d.period, 10);
+    const { type, countBefore, changedIds } = d.types.find(t => t.type === 'Gesetzgebung');
+    const localCount = groups.find(c => c.period === period).types
+      .find(ct => ct.type === type).count;
+    // Append Changed IDs
+    update.concat(changedIds);
+    // Compare Counts Remote & Local
+    if (countBefore > localCount) {
+      // Find remote Procedure Updates
+      const { data: { procedureUpdates } } = await client.query({
+        query: getProcedureUpdates,
+        variables: { period, type },
+      });
+      // Find local Procedure Updates
+      const localProcedureUpdates = await ProcedureModel
+        .find({ period, type }, { procedureId: 1, lastUpdateDate: 1 });
+      // Compare
+      procedureUpdates.forEach((pu) => {
+        const localData = localProcedureUpdates.find(ld => ld.procedureId === pu.procedureId);
+        if (!localData || new Date(localData.lastUpdateDate) < new Date(pu.updatedAt)) {
+          update.push(pu.procedureId);
+        }
+      });
+    }
+  }));
+  // Update
+  return importProcedures(update);
+};
